% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alfalfa.R
\docType{data}
\name{alfalfa}
\alias{alfalfa}
\title{Alfalfa Experiment}
\format{
data.frame
}
\usage{
alfalfa
}
\description{
This experiment was established to compare 12 alfalfa (labeled A-L) varieties (\code{variety}).
A total of 6 plots (\code{block}) per variety were established arranged in a RCB design.
The response variable (\code{resp}) corresponds to yield (tons/acre) at harvest time.
}
\examples{
alfalfa |> head()

}
\references{
source:
}
\keyword{datasets}
