% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASRdata.R
\docType{package}
\name{ASRdata}
\alias{ASRdata}
\title{ASRdata: A package with data used in the ASR packages}
\description{
This package presents a series of datasets used on tests and examples for the
ASR package suite.
}
