% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/besag_elbatan.R
\docType{data}
\name{besag.elbatan}
\alias{besag.elbatan}
\title{Wheat varieties in RCBD}
\format{
data.frame
}
\usage{
besag.elbatan
}
\description{
This randomized complete block design experiment was carried out at
El Batán, Mexico, by the International Center for Improvement of Maize
and Wheat (CIMMYT) for testing wheat varieties (\code{gen}) regarding their
ability to produce seeds (\code{yield}). The trial was carried in a single
location with three single-column (\code{col}) replicates and
fifty rows (\code{row}). A total of 50 treatments (\code{varieties}) were evaluated.
For further information refer to the paper mentioned below.
The dataset is available from the agridat package.
}
\examples{
besag.elbatan |> head()

}
\references{
Kevin Wright (2021). agridat: Agricultural Datasets.
R package version 1.18. <https://CRAN.R-project.org/package=agridat>
}
\keyword{datasets}
