% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_apricot.R
\docType{data}
\name{geno.apricot}
\alias{geno.apricot}
\title{Genotypic data for apricot dataset}
\format{
matrix
}
\usage{
geno.apricot
}
\description{
Genotypic data for a total of 60,991 SNPs (coded as 0, 1, 2 and \code{NA}
for missing) on 155 genotypes of apricot (\emph{Prunus armeniaca}).
Dataset obtained from supplementary material in Nsibi \emph{et al.} (2020).

For more information refer to the original publication.
}
\examples{
geno.apricot[1:5, 1:5]

}
\references{
Nsibi M., Gouble B., Bureau S., Flutre T., Sauvage C.,
Audergon J.M., Regnard J.L. 2020. Adoption and optimization of genomic
selection to sustain breeding for Apricot fruit quality.
\emph{G3 Genes, Genomes, Genet.} 10(12):4513-4529.
}
\keyword{datasets}
