% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_pine655.R
\docType{data}
\name{geno.pine655}
\alias{geno.pine655}
\title{Genotypic data of 655 genotypes for loblolly pine dataset}
\format{
matrix
}
\usage{
geno.pine655
}
\description{
Genotypic data for a total of 4,853 SNPs (coded as 0, 1, 2 and -9 for
missing) on 655 genotypes of Loblolly Pine (\emph{Pinus taeda} L.).
Dataset modified from supplementary material from Resende \emph{et al.}
(2012). This dataset differs from the original as some genotypes were made
artificially missing by full-sib family.
}
\examples{
geno.pine655[1:5, 1:5]

}
\references{
Resende M.F.R., Munoz P., Resende M.D.V., Garrick D.J., Fernando R.L.,
Davis J.M., Jokela E.J., Martin T.A., Peter G.F., and Kirst M. 2012.
Accuracy of genomic selection methods in a standard data set of
loblolly pine (\emph{Pinus taeda} L.). Genetics 190:1503-1510.
}
\keyword{datasets}
