% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_salmon.R
\docType{data}
\name{geno.salmon}
\alias{geno.salmon}
\title{Genotypic data for Atlantic salmon dataset}
\format{
matrix
}
\usage{
geno.salmon
}
\description{
Genotypic data on 1,481 Atlantic salmon samples. A total of 17,156
SNP markers (coded as 0, 1, 2 and \code{NA} for missing) are included in this
dataset. Dataset obtained from supplementary material in Robledo
\emph{et al.} (2018).
}
\details{
For more information refer to the original publication.
}
\examples{
geno.salmon[1:5, 1:5]

}
\references{
Robledo D., Matika O., Hamilton A., and Houston R.D. 2018.
Genome-wide association and genomic selection for resistance
to amoebic gill disease in Atlantic salmon.
G3 Genes, Genomes, Genetics 8:1195-1203.
}
\keyword{datasets}
