% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_sealice.R
\docType{data}
\name{geno.sealice}
\alias{geno.sealice}
\title{Genotypic data for Atlantic salmon challenged with sea lice}
\format{
matrix
}
\usage{
geno.sealice
}
\description{
Genotypic data on 624 (90 parents and 534 offspring) Atlantic
salmon samples challenged with sea lice. A total of 78,362 SNP markers
(coded as 0, 1, 2 and \code{NA} for missing) are included in this dataset.
Dataset obtained from supplementary material in Tsai \emph{et al.} (2017).
}
\details{
Pre-processing by the authors: 1) SNPs with MAF < 5% eliminated;
2) SNPs with unknown position (indels) were removed.

For more information refer to the original publication.
}
\examples{
geno.sealice[1:5, 1:5]

}
\references{
Tsai H.Y., Matika O., McKinnon Edwards S., Antolín–Sánchez R.,Hamilton A.,
Guy D.R., Tinch A.E., Gharbi K., Stear M.J., Taggart J.B., Bron J.E.,
Hickey J.M., Houston R.D. 2017. Genotype imputation to improve the
cost-efficiency of genomic selection in farmed Atlantic salmon.
\emph{G3 Genes, Genomes, Genet.} 6:1313-1326.
}
\keyword{datasets}
