% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafarea.R
\docType{data}
\name{leafarea}
\alias{leafarea}
\title{Disease Resistance}
\format{
data.frame
}
\usage{
leafarea
}
\description{
This trial investigates the resistance of 12 varieties of a plant to a soil borne disease.
The trial was done in a glasshouse based on a randomized complete block design
(RCBD) with 10 blocks (\code{block}). Each of these blocks consisted in 24 pots (\code{pot}), with a single
plant per pot, which had randomly assigned one of the 12 varieties (\code{variety}) and one of the
two types (\code{disease}) of soil: healthy (H) or infected (D). Therefore, we have a 12 x 2 factorial
experiment with the two treatment (\code{trt}) factors: disease and variety. In this study,
the response variable corresponds to total leaf area (\code{leafarea}) of each plant (in cm).
}
\examples{
leafarea |> head()

}
\references{
source:
}
\keyword{datasets}
