% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_apricot.R
\docType{data}
\name{map.apricot}
\alias{map.apricot}
\title{Genomic map for apricot dataset}
\format{
data frame
}
\usage{
map.apricot
}
\description{
Genomic map for a total of 60,991 SNPs with marker names (\code{marker}),
chromosomes (\code{chrom}), and positions (\code{pos}).
Dataset obtained from supplementary material in Nsibi \emph{et al.} (2020).

For more information refer to the original publication.
}
\examples{
map.apricot |> head()

}
\references{
Nsibi M., Gouble B., Bureau S., Flutre T., Sauvage C.,
Audergon J.M., Regnard J.L. 2020. Adoption and optimization of genomic
selection to sustain breeding for Apricot fruit quality.
\emph{G3 Genes, Genomes, Genet.} 10(12):4513-4529.
}
\keyword{datasets}
