% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_sealice.R
\docType{data}
\name{map.sealice}
\alias{map.sealice}
\title{Genomic map for Atlantic salmon challenged with sea lice}
\format{
data frame
}
\usage{
map.sealice
}
\description{
Genomic map for a total of 78,362 SNPs with marker names (\code{marker}),
chromosomes (\code{chrom}), and positions (\code{pos}).
Dataset obtained from supplementary material in Tsai \emph{et al.} (2017).

For more information refer to the original publication.
}
\examples{
map.sealice |> head()

}
\references{
Tsai H.Y., Matika O., McKinnon Edwards S., Antolín–Sánchez R.,Hamilton A.,
Guy D.R., Tinch A.E., Gharbi K., Stear M.J., Taggart J.B., Bron J.E.,
Hickey J.M., Houston R.D. 2017. Genotype imputation to improve the
cost-efficiency of genomic selection in farmed Atlantic salmon.
\emph{G3 Genes, Genomes, Genet.} 6:1313-1326.
}
\keyword{datasets}
