% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped_salmon.R
\docType{data}
\name{ped.salmon}
\alias{ped.salmon}
\title{Pedigree data for Atlantic salmon dataset}
\format{
data.frame
}
\usage{
ped.salmon
}
\description{
Pedigree data of 1,481 Atlantic salmon samples.
Missing parental information coded as 0.
Dataset obtained from supplementary material in Robledo \emph{et al.} (2018).
}
\examples{
ped.salmon |> head()

}
\references{
Robledo D., Matika O., Hamilton A., and Houston R.D. 2018.
Genome-wide association and genomic selection for resistance
to amoebic gill disease in Atlantic salmon.
G3 Genes, Genomes, Genetics 8:1195-1203.
}
\keyword{datasets}
