% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_apple.R
\docType{data}
\name{pheno.apple}
\alias{pheno.apple}
\title{Phenotypic data for apple dataset}
\format{
data.frame
}
\usage{
pheno.apple
}
\description{
Phenotypic data on 247 apple clones (\emph{i.e.}, genotypes) evaluated for
several fruit quality traits at two New Zealand sites, Motueka (MOT) and
Hawkes Bay (HB). Dataset obtained from supplementary material in Kumar
\emph{et al.} (2015).
}
\examples{
pheno.apple |> head()

}
\references{
Kumar S., Molloy C., Muñoz P., Daetwyler H., Chagné D., and Volz R. 2015.
Genome-enabled estimates of additive and nonadditive genetic variances and
prediction of apple phenotypes across environments. G3 Genes, Genomes,
Genetics 5:2711–2718.
}
\keyword{datasets}
