% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_apricot.R
\docType{data}
\name{pheno.apricot}
\alias{pheno.apricot}
\title{Raw phenotypic information for apricot dataset}
\format{
data frame
}
\usage{
pheno.apricot
}
\description{
The plant material consists of a F1 generation from a cross between cultivars
"Goldrich" and "Moniqui". The progeny is composed of 153 individuals
evaluated at the INRAE experimental field in southern France.
Only data for year 2006 is included here.

Fruits were collected and divided into three homogenous (maturity) lots of
four fruits per genotype. Fruits were then measured for weight
(\code{F.Weight}), ground colour (\code{Hue.g}), ethylene (\code{Ethylene}),
refractive index (solid soluble content; \code{RI}), titratable acidity
(\code{TA}), glucose (\code{Glucose}), fructose (\code{Fructose}),
citric acid (\code{Citric.A.}), and malic acid (\code{Malic.A.}).

For more information refer to the original publication.
}
\examples{
pheno.apricot |> head()

}
\references{
Nsibi M., Gouble B., Bureau S., Flutre T., Sauvage C.,
Audergon J.M., Regnard J.L. 2020. Adoption and optimization of genomic
selection to sustain breeding for Apricot fruit quality.
\emph{G3 Genes, Genomes, Genet.} 10(12):4513-4529.
}
\keyword{datasets}
