% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_pine.R
\docType{data}
\name{pheno.pine}
\alias{pheno.pine}
\title{Phenotypic data for loblolly pine dataset}
\format{
data.frame
}
\usage{
pheno.pine
}
\description{
Deregressed estimated breeding values (DEBV) for the trait diameter at breast
height (DBH) at 6 years of age from trees grown at site Nassau.
The dataset contains a total of 861 genotypes of loblolly pine
(\emph{Pinus taeda} L.).
Dataset obtained from supplementary material in Resende \emph{et al.} (2012).
}
\details{
For more information refer to the original publication.
}
\examples{
pheno.pine |> head()

}
\references{
Resende M.F.R., Munoz P., Resende M.D.V., Garrick D.J., Fernando R.L.,
Davis J.M., Jokela E.J., Martin T.A., Peter G.F., and Kirst M. 2012.
Accuracy of genomic selection methods in a standard data set of
loblolly pine (\emph{Pinus taeda} L.). Genetics 190:1503-1510.
}
\keyword{datasets}
