% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_salmon.R
\docType{data}
\name{pheno.salmon}
\alias{pheno.salmon}
\title{Phenotypic data for Atlantic salmon dataset}
\format{
data.frame
}
\usage{
pheno.salmon
}
\description{
Phenotypic data on 1,481 Atlantic salmon individuals.
All fish were phenotyped for mean gill score (mean of the left
gill and right gill scores) and amoebic load (qPCR values using
\emph{Neoparamoeba perurans} specific primers, amplified from one of the
gills). Dataset obtained from supplementary material in Robledo \emph{et al.}
(2018).
}
\examples{
pheno.salmon |> head()

}
\references{
Robledo D., Matika O., Hamilton A., and Houston R.D. 2018.
Genome-wide association and genomic selection for resistance
to amoebic gill disease in Atlantic salmon.
G3 Genes, Genomes, Genetics 8:1195-1203.
}
\keyword{datasets}
