% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_sealice.R
\docType{data}
\name{pheno.sealice}
\alias{pheno.sealice}
\title{Phenotypic data for Atlantic salmon challenged with sea lice}
\format{
data.frame
}
\usage{
pheno.sealice
}
\description{
Phenotypic data on 624 Atlantic salmon of breeding program.
Genotypes correspond to 59 nuclear families from 90 parents.
One year old fish were challenged with sea lice (L. salmonis).
After seven days, the fish were measured for body weight (g; \code{weight})
and number of lice attached to the fish (\code{lice}).
Dataset obtained from supplementary material in Tsai \emph{et al.} (2017).
}
\details{
For more information refer to the original publication.
}
\examples{
pheno.sealice |> head()

}
\references{
Tsai H.Y., Matika O., McKinnon Edwards S., Antolín–Sánchez R.,Hamilton A.,
Guy D.R., Tinch A.E., Gharbi K., Stear M.J., Taggart J.B., Bron J.E.,
Hickey J.M., Houston R.D. 2017. Genotype imputation to improve the
cost-efficiency of genomic selection in farmed Atlantic salmon.
\emph{G3 Genes, Genomes, Genet.} 6:1313-1326.
}
\keyword{datasets}
