% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_trout.R
\docType{data}
\name{pheno.trout}
\alias{pheno.trout}
\title{Phenotypic data for rainbow trout dataset}
\format{
data.frame
}
\usage{
pheno.trout
}
\description{
Phenotypic information for a total of 749 rainbow trouts
(Oncorhynchus mykiss) from 58 full-sib families. The fish were challenged
with Infectious pancreatic necrosis virus (IPNV). Records includes the
identification of the fish (\code{ID}), family (\code{family}),
and mortality (\code{ipn_mortality}). Dataset obtained from supplementary
material in Rodríguez \emph{et al.} (2019).
}
\details{
For more information refer to the original publication.
}
\examples{
pheno.trout |> head()

}
\references{
Rodríguez F.H., Flores-Mara R., Yoshida G.M., Barría A., Jedlicki A.M.,
Lhorente J.P., Reyes-López F., Yáñez J.M. 2019. Genome-Wide Association
Analysis for resistance to infectious pancreatic necrosis virus identifies
candidate genes involved in viral replication and immune response in rainbow
trout (\emph{Oncorhynchus mykiss}). G3 (Bethesda) 9(9):2897-2904.
doi: 10.1534/g3.119.400463.
}
\keyword{datasets}
