% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel.R
\docType{data}
\name{pixel}
\alias{pixel}
\title{Repeated CT scans on dogs}
\format{
data.frame
}
\usage{
pixel
}
\description{
This dataset was collected by performing CT scans on dogs (\code{dog}).
Scans of the right and left (\code{side}) lymph nodes in the axillary region
were recorded in form of pixel intensity (\code{pixel}). A total of 10 dogs
were scanned over a period of 14 days. The dogs had received an intravenous
application of a dye contrast at time zero.
}
\examples{
pixel |> head()

}
\references{
Pinheiro & Bates (2000, p.42 and p.440). Mixed Effects Models in S and S-plus.
}
\keyword{datasets}
