% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rail.R
\docType{data}
\name{rail}
\alias{rail}
\title{Stress in railways rails}
\format{
data.frame
}
\usage{
rail
}
\description{
This dataset originates from a nondestructive experiment that tested for
stress in railways rails. The sampling consisted in randomly choosing six
rails (\code{rail}) and testing each three times by measuring the time it
takes for a given type of ultrasonic wave to travel the complete length of
the rail (\code{travel}).
}
\examples{
rail |> head()

}
\references{
Pinheiro & Bates (2000, p.4 and p.443). Mixed Effects Models in S and S-plus.
Originally cited by Devore (2000, p. 427).
}
\keyword{datasets}
