% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratpup.R
\docType{data}
\name{ratpup}
\alias{ratpup}
\title{Rat pup weight}
\format{
data.frame
}
\usage{
ratpup
}
\description{
This dataset contains the birth weights of pups (\code{weight}) from litters
(\code{litter}) born to female rats that received different doses of an
experimental compound. The treatments (\code{treatment}) corresponded to the
doses: high, low and control. The data consisted originally of 30 female
rats with litter sizes (\code{lsize}) that vary widely. The variable
\code{sex} identifies the corresponding sex of each pup.
}
\examples{
ratpup |> head()

}
\references{
Pinheiro & Bates (2000, p.424). Mixed Effects Models in S and S-plus.
}
\keyword{datasets}
