% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semicond.R
\docType{data}
\name{semicond}
\alias{semicond}
\title{Oxide layers on silicon wafers}
\format{
data.frame
}
\usage{
semicond
}
\description{
In this dataset, the measurements correspond to thickness (\code{thick}) of
the oxide layer on silicon wafers. The wafers (\code{wafer}) were randomly
selected from eight different lots (\code{lot}), each having 25 wafers in
total. After the layer of oxide was deposited, the layer’s thickness was
assessed at 3 randomly selected sites (\code{site}) within each wafer.
In addition, two different sources (\code{source}) of material are available
and can be compared statistically.
}
\examples{
semicond |> head()

}
\references{
Littell et al. (2006, p.81). SAS for Linear Mixed Models, 2nd Edition.
}
\keyword{datasets}
