% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinlong.R
\docType{data}
\name{twinlong}
\alias{twinlong}
\title{IQ scores of twins}
\format{
data.frame
}
\usage{
twinlong
}
\description{
This dataset contains the IQ scores (\code{iq}) of identical
(\code{monozygotic}) twins (\code{twin}) that were separated early in life
and reared apart. Univitelline twins are very important for genetic studies
because they share virtually identical genotype. The data corresponds to a
total of 32 pairs (\code{pair}) of twins where one of the members (twin = A)
was reared by their natural parent, and the other (twin = B) was reared by a
relative or other person.
}
\examples{
twinlong |> head()

}
\references{
McClave & Sincich (2009, p.473). Statistics 11th Edition.
}
\keyword{datasets}
